Add from_shape and deprecate from_dim_slice for ArrayView, ArrayViewMut
