+++ src/impl_constructors.rs
@@ -15,105 +15,10 @@
 use libnum::{Zero, One, Float};
 
 use imp_prelude::*;
+use {Shape, StrideShape};
 use dimension;
 use linspace;
 use error::{self, ShapeError, ErrorKind};
-
-/// A contiguous array shape of n dimensions.
-///
-/// Either c- or f- memory ordered (*c* a.k.a *row major* is the default).
-#[derive(Copy, Clone, Debug)]
-pub struct Shape<D> {
-    dim: D,
-    is_c: bool,
-}
-
-/// An array shape of n dimensions c-order, f-order or custom strides.
-#[derive(Copy, Clone, Debug)]
-pub struct StrideShape<D> {
-    dim: D,
-    strides: D,
-    custom: bool,
-}
-
-/// A trait for `Shape` and `D where D: Dimension` that allows
-/// customizing the memory layout (strides) of an array shape.
-pub trait ShapeBuilder {
-    type Dim: Dimension;
-
-    fn f(self) -> Shape<Self::Dim>;
-    fn set_f(self, is_f: bool) -> Shape<Self::Dim>;
-    fn strides(self, strides: Self::Dim) -> StrideShape<Self::Dim>;
-}
-
-impl<D> From<D> for Shape<D>
-    where D: Dimension
-{
-    fn from(d: D) -> Self {
-        Shape {
-            dim: d,
-            is_c: true,
-        }
-    }
-}
-
-impl<D> From<D> for StrideShape<D>
-    where D: Dimension
-{
-    fn from(d: D) -> Self {
-        StrideShape {
-            strides: d.default_strides(),
-            dim: d,
-            custom: false,
-        }
-    }
-}
-
-impl<D> From<Shape<D>> for StrideShape<D>
-    where D: Dimension
-{
-    fn from(shape: Shape<D>) -> Self {
-        let d = shape.dim;
-        let st = if shape.is_c { d.default_strides() } else { d.fortran_strides() };
-        StrideShape {
-            strides: st,
-            dim: d,
-            custom: false,
-        }
-    }
-}
-
-impl<D> ShapeBuilder for D
-    where D: Dimension
-{
-    type Dim = D;
-    fn f(self) -> Shape<D> { self.set_f(true) }
-    fn set_f(self, is_f: bool) -> Shape<D> {
-        Shape::from(self).set_f(is_f)
-    }
-    fn strides(self, st: D) -> StrideShape<D> {
-        Shape::from(self).strides(st)
-    }
-}
-
-impl<D> ShapeBuilder for Shape<D>
-    where D: Dimension
-{
-    type Dim = D;
-    fn f(self) -> Self { self.set_f(true) }
-    fn set_f(mut self, is_f: bool) -> Self {
-        self.is_c = !is_f;
-        self
-    }
-    fn strides(self, st: D) -> StrideShape<D> {
-        StrideShape {
-            dim: self.dim,
-            strides: st,
-            custom: true,
-        }
-    }
-}
-
 
 /// Constructor methods for one-dimensional arrays.
 ///
